Simplify example and remove the launcher UI
